part2

+++ src/app/module/patient/patient.service.ts
@@ -1,5 +1,7 @@
 import { Prisma, Patient } from "../../../generated/prisma/client";
+import AppError from "../../errorHelpers/AppError";
 import { prisma } from "../../lib/prisma";
+import status from "http-status";
 
 // CREATE
 const createPatients = async (
@@ -7,20 +9,31 @@ const createPatients = async (
 ) => {
   const result = await prisma.patient.createMany({
     data: payload,
-    skipDuplicates: true, // একই email থাকলে skip করবে
+    skipDuplicates: true, // duplicate email থাকলে skip করবে
   });
-  return result; // শুধু inserted count দিবে
+
+  return result;
 };
+
 // GET ALL
 const getAllPatients = async (): Promise<Patient[]> => {
-  return await prisma.patient.findMany();
+  const result = await prisma.patient.findMany();
+  return result;
 };
 
 // UPDATE
 const updatePatient = async (
   id: string,
   payload: Partial<Prisma.PatientUpdateInput>
-) => {
+): Promise<Patient> => {
+  const patient = await prisma.patient.findUnique({
+    where: { id },
+  });
+
+  if (!patient) {
+    throw new AppError(status.NOT_FOUND, "Patient not found");
+  }
+
   const result = await prisma.patient.update({
     where: { id },
     data: payload,
@@ -31,16 +44,20 @@ const updatePatient = async (
 
 // DELETE
 const deletePatient = async (id: string): Promise<Patient> => {
-  try {
-    const result = await prisma.patient.delete({
-      where: { id },
-    });
-    return result;
-  } catch (error) {
-    throw new Error("Patient not found or already deleted", { cause: error });
-  }
-};
+  const patient = await prisma.patient.findUnique({
+    where: { id },
+  });
 
+  if (!patient) {
+    throw new AppError(status.NOT_FOUND, "Patient not found or already deleted");
+  }
+
+  const result = await prisma.patient.delete({
+    where: { id },
+  });
+
+  return result;
+};
 
 export const PatientService = {
   createPatients,
